Moved around constants and base url
