+++ config/constants.ts
@@ -1,5 +1,13 @@
 import { ServiceAccount } from "next-firebase-auth-edge/lib/auth/credential";
 
+
+/**
+ * The base url for the server side code
+ * https://github.com/vercel/next.js/discussions/16429
+ * Defaults to vercel url, will fallback to env variable for local host
+ */
+export const BaseURL = process.env.NEXT_PUBLIC_VERCEL_URL ?? process.env.NEXT_PUBLIC_SITE_URL ?? "http://localhost:3000/";
+
 /**
  * Define the firebase service account credentials for firebase admin
  */
